+++ src/pages/JobDescription/components/JobDescriptionItem.jsx
@@ -1,10 +1,10 @@
 import { useTheme } from "@emotion/react";
 import {
-  ToggleButton,
   Grid,
-  ToggleButtonGroup,
   ListItem,
   ListItemText,
+  ToggleButton,
+  ToggleButtonGroup,
 } from "@mui/material";
 import PropTypes from "prop-types";
 import React from "react";
