+++ src/routes.jsx
@@ -5,7 +5,7 @@ import CreateJobProfilePage from "./pages/CreateJobProfile/CreateJobProfilePage"
 import DashboardPage from "./pages/Dashboard/DashboardPage";
 import EmployeeOverviewPage from "./pages/EmployeeOverview/EmployeeOverviewPage";
 import HomePage from "./pages/Home/HomePage";
-import JobDescriptionPage from "./pages/JobDescribtion/JobDescriptionPage";
+import JobDescriptionPage from "./pages/JobDescription/JobDescriptionPage";
 import JobProfilePage from "./pages/JobProfile/JobProfilePage";
 import LinkedInLearning from "./pages/LinkedInLearning/LinkedInLearningPage";
 import LoginPage from "./pages/Login/LoginPage";
